Implement "Step 2.2 Model Architecture Design"
The model architecture at this stage is build up by fully connected 'dense' layers and concolutional layers. 
In case of the conv layers, the creating of mulitple layers can be facilitated by looping over the desired number of conv layer. The naming conflict are solved by using 'tf.vairable_scope' with dynamic names.
This is an early and unstable version.