Update "Step 2.2" with inception-v4 Block-B
Implemented the inceptions-v4 Block-B and added it to the model as a test. Worked fine.
The Block-B is design according to the original paper for 17x17 grids, so it fits well as one of the middle layers in the model with an 16x16 grid.
The filter sizes had to be far smaller, due to limited computational power,  than suggested in the paper. This should be changed when training the model on a powerful computer.