Update "Step 2.2" with inception-v4 Block-C
Implemented the inceptions-v4 Block-C and added it to the model as a test. Worked fine.
The Block-C is design according to the original paper for 8x8 grids, so it fits well as one of the last layers in the model with an 8x8 grid.
The filter sizes had to be far smaller, due to limited computational power,  than suggested in the paper. This should be changed when training the model on a powerful computer.